fix: compiles (commenting out tests)

+++ src/Catalogue.sol
@@ -63,12 +63,12 @@ contract Catalogue {
         Auctioneer.Routing memory routing = getRouting(lotId_);
 
         // Calculate fees
-        (uint96 protocolFee, uint96 referrerFee) = FeeManager(auctionHouse).calculateQuoteFees(
+        (uint256 protocolFee, uint256 referrerFee) = FeeManager(auctionHouse).calculateQuoteFees(
             keycodeFromVeecode(routing.auctionReference), true, amount_
         ); // we assume there is a referrer to give a conservative amount
 
         // Get payout from module
-        return module.payoutFor(lotId_, amount_ - protocolFee - referrerFee);
+        return module.payoutFor(lotId_, amount_ - uint96(protocolFee) - uint96(referrerFee));
     }
 
     function priceFor(uint96 lotId_, uint96 payout_) external view returns (uint256) {
